app/requirements.txt
fastapi
celery
uvicorn
flower
python-dotenv
pyamqp
uvicorn